requirements.txt
Flask
python-dotenv
openai
PyMuPDF
fpdf
